Initial commit for the rust implementation of prolly tree
A Prolly Tree is a hybrid data structure that combines the features of B-trees and
Merkle trees to provide both efficient data access and verifiable integrity.

It is specifically designed to handle the requirements of distributed systems and
large-scale databases, making indexes syncable and distributable over peer-to-peer (P2P) networks.